Change int wrap to math.trunc and add import math
I was testing int() at first thinking it behaved the same as it does in Java, but Int() in Python rounds so I decided to use math.trunc. Forgot to change that one line though, thanks for pointing it out. Also forgot to add import math to this file. Not 100% sure it matters as the add function was working without math.trunc anyway but better safe than sorry.